requirements.txt
Django>=4.1.6
mysqlclient>=2.1
numpy>=1.24.2
matplotlib>=3.6.3
paho-mqtt>=1.6.1
mysql-connector-python>=8.0.32
redis>=4.5.1
hiredis>=2.2.2
djangorestframework~=3.14.0
scipy~=1.10.1
Pillow~=9.4.0